share code across rules
